collector/requirements.txt
requests==2.32.3
pymongo==4.8.0
python-dotenv==1.0.1
